Large code cleanup Removes some unecessary files Renames Educator Landing to MainLanding Abstracts some code into seperate files
