+++ client/src/components/App.jsx
@@ -2,7 +2,6 @@ import React from 'react';
 
 export default React.createClass({
   render: function() {
-    console.log('In App.jsx', this.props.children);
     return this.props.children;
   }
 });
