+++ client/src/reduxStateEmitterMiddleware.js
@@ -1,9 +1,6 @@
 export default socket => store => next => action => {
-  console.log('in middleware', action);
-  if (action.meta && action.meta.remote) { // so SET_STATE infinite loop isn't triggered and is only set on cleint side
-    console.log('emitted action');
+  if (action.meta && action.meta.remote) { // so SET_STATE infinite loop isn't triggered and is only set on client side
     socket.emit('action', action);
-
   }
   return next(action);
 };
