--- server/src/store.js
@@ -1,6 +1,0 @@
-import {createStore} from 'redux';
-import reducer from './reducer';
-
-export default function makeStore() {
-  return createStore(reducer);
-}
